Expression split, advanced quantified expression
The expression slides are too large and had some sort of a weird intro
at the right point in the code, this moves most of the complexity onto
the dedicated chapters.

eg: subtypes is now a proper section in basic types.

As for quantified expressions, this creates a new chapter: this did go
into too many details for people that would not write Ada 2012, and as
such this does not need a "default" inclusion.