Overview: Moved "strong & static type" to decl.
Overview is very large, and it was getting into lots of details (4
slides) on the type system *before* we get to actually talk about it.

As a result, we had to repeat some of that info where it made more
sense: in the type system slides.

This commit splits that up, by moving the information from overview to
declarations and basic type chapters.