Overview: Simplify packages slides
Use the same "spec /= impl" as for the subprogram slide above.

Remove the slide on views, this is far too advanced at this point and
the slide just above covers encapsulation.

Move the ADT definition to basic type slide where it makes tons of sense
to be (basically we would re-intro them at this point). Maybe the two
resulting slides could end up being merged?

OOP was referencing ADT, that's too advanced at this point too (people
haven't written a single line of code yet, it's an overview!)