Add module locations to TEXINPUTS environment variable
When trying to build a multi-course slide deck, some
(or all) of the modules are going to be in a different
directory than the tool is run from. So we want to add
all of those directories to TEXINPUTS by parsing the
source file and determining absolute paths.

This commit does NOT fix the problem because Pandoc
doesn't deal with relative paths well. We are already
fixing this for the "source_include" container, now we need
to do it for the "include" directive